+++ packages/website/components/contexts/tokensContext.js
@@ -13,6 +13,8 @@ import { deleteToken, getTokens, createToken } from 'lib/api';
  * @property {boolean} isFetchingTokens Whether or not new Tokens are being fetched
  * @property {boolean} isCreating Whether or not a new token is being created
  * @property {number|undefined} fetchDate The date in which the last Tokens list fetch happened
+ * @property {string|undefined} errorMessage The error message if any
+ * @property {boolean} hasError Whether or not there was an error
  */
 
 /**
@@ -35,30 +37,69 @@ export const TokensProvider = ({ children }) => {
   const [isFetchingTokens, setIsFetchingTokens] = useState(false);
   const [isCreating, setIsCreating] = useState(false);
   const [fetchDate, setFetchDate] = useState(/** @type {number|undefined} */ (undefined));
+  const [hasError, setHasError] = useState(/** @type {boolean} */ (false));
+  const [errorMessage, setErrorMessage] = useState(/** @type {string} */ (''));
+
+  const processApiError = useCallback(
+    error => {
+      if (!error.message) {
+        setHasError(true);
+        setErrorMessage('An unknown error has occurred');
+        return;
+      }
+      const errorObject = JSON.parse(
+        error.message.substring(error.message.indexOf('{'), error.message.lastIndexOf('}') + 1)
+      );
+      setHasError(true);
+      if (errorObject.code === 'ERROR_MAINTENANCE') {
+        setErrorMessage('Tokens API undergoing maintenance. Please try again later.');
+        return;
+      }
+
+      setErrorMessage('Tokens API error has occurred. Please try again later.');
+    },
+    [setErrorMessage, setHasError]
+  );
 
   const getTokensCallback = useCallback(
     /** @type {() => Promise<Token[]>}} */
     async () => {
-      setIsFetchingTokens(true);
-      const updatedTokens = await getTokens();
-      setTokens(updatedTokens);
-      setIsFetchingTokens(false);
-      setFetchDate(Date.now());
+      try {
+        setIsFetchingTokens(true);
+        const updatedTokens = await getTokens();
+        setTokens(updatedTokens);
+        setIsFetchingTokens(false);
+        setFetchDate(Date.now());
 
-      return updatedTokens;
+        return updatedTokens;
+      } catch (e) {
+        processApiError(e);
+
+        return new Promise((resolve, reject) => {
+          reject(errorMessage);
+        });
+      }
     },
-    [setTokens, setFetchDate, setIsFetchingTokens]
+    [setTokens, setFetchDate, setIsFetchingTokens, processApiError, errorMessage]
   );
 
   const createTokensCallback = useCallback(
     /** @type {(name: string) => Promise<Token>}} */
     async name => {
-      setIsCreating(true);
-      const newToken = await createToken(name);
-      setIsCreating(false);
-      return newToken;
+      try {
+        setIsCreating(true);
+        const newToken = await createToken(name);
+        setIsCreating(false);
+        return newToken;
+      } catch (e) {
+        processApiError(e);
+
+        return new Promise((resolve, reject) => {
+          reject(errorMessage);
+        });
+      }
     },
-    [setIsCreating]
+    [setIsCreating, processApiError, errorMessage]
   );
 
   return (
@@ -66,6 +107,8 @@ export const TokensProvider = ({ children }) => {
       value={
         /** @type {TokensContextProps} */
         ({
+          hasError,
+          errorMessage,
           deleteToken,
           createToken: createTokensCallback,
           getTokens: getTokensCallback,