+++ packages/website/components/footer/footer.js
@@ -4,11 +4,12 @@ import { useRouter } from 'next/router';
 import Link from 'next/link';
 import clsx from 'clsx';
 
+import { trackCustomLinkClick, events } from '../../lib/countly';
 import SiteLogo from '../../assets/icons/w3storage-logo.js';
 import Button from '../button/button';
-import Squiggle from '../../assets/illustrations/squiggle.js';
+import Img from '../cloudflareImage.js';
+import ImageSquiggle from '../../public/images/illustrations/squiggle.png';
 import GeneralPageData from '../../content/pages/general.json';
-import { trackCustomLinkClick, events } from 'lib/countly';
 
 // ====================================================================== Params
 /**
@@ -51,7 +52,9 @@ export default function Footer({ isProductApp }) {
   return (
     <footer id="site-footer" className={clsx(isProductApp ? 'clear-bg' : '')}>
       <section id="site-footer-section">
-        <Squiggle id="footer_squiggle" />
+        <div id="footer_squiggle">
+          <Img alt="" src={ImageSquiggle} />
+        </div>
 
         <div className="grid">
           <div className="col-4_sm-8_mi-12" data-push-left="off-1_sm-1_mi-0">