+++ packages/website/declaration.d.ts
@@ -1,3 +1,6 @@
 declare module '*.css';
 declare module '*.scss';
 declare module '*.svg';
+declare module '*.png';
+declare module '*.jpg';
+declare module '*.gif';