+++ packages/website/pages/account/index.js
@@ -79,13 +79,13 @@ const Account = () => {
   return (
     <>
       <div className="page-container account-container">
-        <h1 className="account-heading">{dashboard.heading}</h1>
+        <h1 className="table-heading">{dashboard.heading}</h1>
         <div className="account-content">
           <StorageManager content={AppData.page_content.storage_manager} className="account-storage-manager" />
           <CTACard
             className="account-upload-cta"
             {...CTAConfigs[CTACardTypes.UPLOAD_FILES]}
-            background={<GradientBackground variant="upload-cta-gradient" />}
+            background={<GradientBackground variant="upload-cta-gradient-small" />}
           />
           <FilesManager
             content={AppData.page_content.file_manager}